requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
rich==13.7.0
colorama==0.4.6
python-dotenv==1.0.0
python-telegram-bot==20.7

